+++ internal/api/hvac.go
@@ -7,6 +7,7 @@ import (
 	"strconv"
 
 	"github.com/energieip/common-components-go/pkg/dhvac"
+	"github.com/energieip/common-components-go/pkg/duser"
 
 	"github.com/energieip/srv200-coreservice-go/internal/core"
 	"github.com/energieip/srv200-coreservice-go/internal/database"
@@ -17,31 +18,35 @@ import (
 func (api *API) readHvacConfig(w http.ResponseWriter, mac string) {
 	driver, _ := database.GetHvacConfig(api.db, mac)
 	if driver == nil {
-		api.sendError(w, APIErrorDeviceNotFound, "Device "+mac+" not found")
+		api.sendError(w, APIErrorDeviceNotFound, "Device "+mac+" not found", http.StatusInternalServerError)
 		return
 	}
-
-	inrec, _ := json.MarshalIndent(driver, "", "  ")
-	w.Write(inrec)
+	json.NewEncoder(w).Encode(driver)
 }
 
 func (api *API) getHvacSetup(w http.ResponseWriter, req *http.Request) {
-	api.setDefaultHeader(w)
+	if api.hasAccessMode(w, req, []string{duser.PriviledgeAdmin}) != nil {
+		api.sendError(w, APIErrorUnauthorized, "Unauthorized Access", http.StatusUnauthorized)
+		return
+	}
 	params := mux.Vars(req)
 	api.readHvacConfig(w, params["mac"])
 }
 
 func (api *API) setHvacSetup(w http.ResponseWriter, req *http.Request) {
-	api.setDefaultHeader(w)
+	if api.hasAccessMode(w, req, []string{duser.PriviledgeAdmin}) != nil {
+		api.sendError(w, APIErrorUnauthorized, "Unauthorized Access", http.StatusUnauthorized)
+		return
+	}
 	body, err := ioutil.ReadAll(req.Body)
 	if err != nil {
-		api.sendError(w, APIErrorBodyParsing, "Error reading request body")
+		api.sendError(w, APIErrorBodyParsing, "Error reading request body", http.StatusInternalServerError)
 		return
 	}
 	setup := dhvac.HvacSetup{}
 	err = json.Unmarshal(body, &setup)
 	if err != nil {
-		api.sendError(w, APIErrorBodyParsing, "Could not parse input format "+err.Error())
+		api.sendError(w, APIErrorBodyParsing, "Could not parse input format "+err.Error(), http.StatusInternalServerError)
 		return
 	}
 
@@ -51,28 +56,31 @@ func (api *API) setHvacSetup(w http.ResponseWriter, req *http.Request) {
 }
 
 func (api *API) setHvacConfig(w http.ResponseWriter, req *http.Request) {
-	api.setDefaultHeader(w)
+	if api.hasAccessMode(w, req, []string{duser.PriviledgeAdmin, duser.PriviledgeMaintainer}) != nil {
+		api.sendError(w, APIErrorUnauthorized, "Unauthorized Access", http.StatusUnauthorized)
+		return
+	}
 	body, err := ioutil.ReadAll(req.Body)
 	if err != nil {
-		api.sendError(w, APIErrorBodyParsing, "Error reading request body")
+		api.sendError(w, APIErrorBodyParsing, "Error reading request body", http.StatusInternalServerError)
 		return
 	}
 
 	cfg := dhvac.HvacConf{}
 	err = json.Unmarshal(body, &cfg)
 	if err != nil {
-		api.sendError(w, APIErrorBodyParsing, "Could not parse input format "+err.Error())
+		api.sendError(w, APIErrorBodyParsing, "Could not parse input format "+err.Error(), http.StatusInternalServerError)
 		return
 	}
 
 	if cfg.Group != nil {
 		if *cfg.Group < 0 {
-			api.sendError(w, APIErrorInvalidValue, "Invalid groupID "+strconv.Itoa(*cfg.Group))
+			api.sendError(w, APIErrorInvalidValue, "Invalid groupID "+strconv.Itoa(*cfg.Group), http.StatusInternalServerError)
 			return
 		}
 		gr, _ := database.GetGroupConfig(api.db, *cfg.Group)
 		if gr == nil {
-			api.sendError(w, APIErrorDeviceNotFound, "Group "+strconv.Itoa(*cfg.Group)+" not found")
+			api.sendError(w, APIErrorDeviceNotFound, "Group "+strconv.Itoa(*cfg.Group)+" not found", http.StatusInternalServerError)
 			return
 		}
 	}
@@ -83,17 +91,30 @@ func (api *API) setHvacConfig(w http.ResponseWriter, req *http.Request) {
 }
 
 func (api *API) sendHvacCommand(w http.ResponseWriter, req *http.Request) {
-	api.setDefaultHeader(w)
 	body, err := ioutil.ReadAll(req.Body)
 	if err != nil {
-		api.sendError(w, APIErrorBodyParsing, "Error reading request body")
+		api.sendError(w, APIErrorBodyParsing, "Error reading request body", http.StatusInternalServerError)
 		return
 	}
 
 	cmd := core.HvacCmd{}
 	err = json.Unmarshal([]byte(body), &cmd)
 	if err != nil {
-		api.sendError(w, APIErrorBodyParsing, "Could not parse input format "+err.Error())
+		api.sendError(w, APIErrorBodyParsing, "Could not parse input format "+err.Error(), http.StatusInternalServerError)
+		return
+	}
+	dr, _ := database.GetHvacConfig(api.db, cmd.Mac)
+	if dr == nil {
+		api.sendError(w, APIErrorDeviceNotFound, "Device "+cmd.Mac+" not found", http.StatusInternalServerError)
+		return
+	}
+
+	if dr.Group == nil {
+		group := 0
+		dr.Group = &group
+	}
+	if api.hasEnoughRight(w, req, *dr.Group) != nil {
+		api.sendError(w, APIErrorUnauthorized, "Unauthorized Access", http.StatusUnauthorized)
 		return
 	}
 	rlog.Info("Received Hvac cmd", cmd)
@@ -104,25 +125,31 @@ func (api *API) sendHvacCommand(w http.ResponseWriter, req *http.Request) {
 }
 
 func (api *API) getHvacStatus(w http.ResponseWriter, req *http.Request) {
-	api.setDefaultHeader(w)
 	params := mux.Vars(req)
 	mac := params["mac"]
-	led := database.GetHvacStatus(api.db, mac)
-	if led == nil {
-		api.sendError(w, APIErrorDeviceNotFound, "Device "+mac+" not found")
+	dr := database.GetHvacStatus(api.db, mac)
+	if dr == nil {
+		api.sendError(w, APIErrorDeviceNotFound, "Device "+mac+" not found", http.StatusInternalServerError)
 		return
 	}
-	inrec, _ := json.MarshalIndent(led, "", "  ")
-	w.Write(inrec)
+
+	if api.hasEnoughRight(w, req, dr.Group) != nil {
+		api.sendError(w, APIErrorUnauthorized, "Unauthorized Access", http.StatusUnauthorized)
+		return
+	}
+	json.NewEncoder(w).Encode(dr)
 }
 
 func (api *API) removeHvacSetup(w http.ResponseWriter, req *http.Request) {
-	api.setDefaultHeader(w)
+	if api.hasAccessMode(w, req, []string{duser.PriviledgeAdmin}) != nil {
+		api.sendError(w, APIErrorUnauthorized, "Unauthorized Access", http.StatusUnauthorized)
+		return
+	}
 	params := mux.Vars(req)
 	mac := params["mac"]
 	res := database.RemoveHvacConfig(api.db, mac)
 	if res != nil {
-		api.sendError(w, APIErrorDeviceNotFound, "Device "+mac+" not found")
+		api.sendError(w, APIErrorDeviceNotFound, "Device "+mac+" not found", http.StatusInternalServerError)
 		return
 	}
 	w.Write([]byte("{}"))
